+++ lib/disco_app/configuration.rb
@@ -5,6 +5,9 @@ module DiscoApp
     # Required configuration.
     attr_accessor :app_name
 
+    # Set the below if using an application proxy.
+    attr_accessor :app_proxy_prefix
+
     # Optional configuration, usually useful for development environments.
     attr_accessor :skip_proxy_verification
     alias_method  :skip_proxy_verification?, :skip_proxy_verification
@@ -12,12 +15,6 @@ module DiscoApp
     alias_method  :skip_webhook_verification?, :skip_webhook_verification
     attr_accessor :skip_carrier_request_verification
     alias_method  :skip_carrier_request_verification?, :skip_carrier_request_verification
-
-    def initialize
-      @skip_proxy_verification = false
-      @skip_webhook_verification = false
-      @skip_carrier_request_verification = false
-    end
 
   end